move some functionality from fastspecfit to io
